Add TTL vars for CloudFront
This adds variables to configure the TTL settings for the CloudFront
distribution default behavior and sets common defaults.

Additionally, the pre-commit config is updated to enable the Terraform
hooks. This wasn't working with pre-commit.ci because Terraform wasn't
available, but the use of pre-commit.ci for this module is disabled for
now.